fix typo where nodes.position was an integer instead of a float
git-svn-id: svn://svn.zenadmin.org/zena/trunk@1086 aaa19a3b-e121-0410-9d45-962fd44e537e

+++ db/migrate/000_zena/039_default_position.rb
@@ -1,10 +1,10 @@
 class DefaultPosition < ActiveRecord::Migration
   def self.up
-    change_column :nodes, :position, :integer, :default => 0.0
+    change_column :nodes, :position, :float, :default => 0.0
     execute "UPDATE nodes SET position = 0.0 where position = 1.0;"
   end
 
   def self.down
-    change_column :nodes, :position, :integer, :default => 1.0
+    change_column :nodes, :position, :float, :default => 1.0
   end
 end
